change type of vec2_ops from float tou double, better norme

+++ includes/wolf3d.h
@@ -6,7 +6,7 @@
 /*   By: awoimbee <awoimbee@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/14 09:48:50 by wta               #+#    #+#             */
-/*   Updated: 2019/01/16 13:10:25 by awoimbee         ###   ########.fr       */
+/*   Updated: 2019/01/16 13:42:57 by awoimbee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,13 +73,13 @@ t_vec2	rotate2d(t_vec2 vector, const double theta);
 t_vec2	vec2_add(t_vec2 a, const t_vec2 b);
 t_vec2	vec2_sub(t_vec2 a, const t_vec2 b);
 t_vec2	vec2_multv(t_vec2 a, const t_vec2 b);
-t_vec2	vec2_multf(t_vec2 a, const float b);
+t_vec2	vec2_multf(t_vec2 a, const double b);
 t_vec2	vec2_div(t_vec2 a, const t_vec2 b);
-t_vec2	vec2_divf(t_vec2 a, const float b);
-float	vec2_dot(const t_vec2 a, const t_vec2 b);
-float	vec2_mod(const t_vec2 a);
+t_vec2	vec2_divf(t_vec2 a, const double b);
+double	vec2_dot(const t_vec2 a, const t_vec2 b);
+double	vec2_mod(const t_vec2 a);
 t_vec2	vec2_normalize(t_vec2 a);
-float	points_dist(const t_vec2 p1, const t_vec2 p2);
+double	points_dist(const t_vec2 p1, const t_vec2 p2);
 
 void	raycasting(t_info *info);
